requirements.txt
Pillow>=9.0.0
PyInstaller>=5.0.0
tkinterdnd2==0.4.3